Move `row` and `col` properties and `__repr__` method of `BaseCell` from `gspread.base.BaseCell` to `gspread.models.Cell`; delete `gspread.base.BaseCell`
