--- gspread/base.py
@@ -1,19 +1,0 @@
-class BaseCell(object):
-    def __init__(self, worksheet, element):
-        pass
-
-    @property
-    def row(self):
-        """Row number of the cell."""
-        return self._row
-
-    @property
-    def col(self):
-        """Column number of the cell."""
-        return self._col
-
-    def __repr__(self):
-        return '<%s R%sC%s %s>' % (self.__class__.__name__,
-                                   self.row,
-                                   self.col,
-                                   repr(self.value))
